Add `has_else` field to the If_Flow struct
Previously, if an If_Flow did not have an Else portion,
it would still attempt to run an Else_Block.
The reason the ELSE block of the If_Flow would not run
was because the default value of start and end lines were 0.
Now the Else block will only be run if the If_Flow `has_else`